+++ src/Assert.php
@@ -26,6 +26,7 @@ use Traversable;
  * @method static void nullOrNumeric($value, $message = '')
  * @method static void nullOrBoolean($value, $message = '')
  * @method static void nullOrScalar($value, $message = '')
+ * @method static void nullOrObject($value, $message = '')
  * @method static void nullOrResource($value, $type = null, $message = '')
  * @method static void nullOrIsCallable($value, $message = '')
  * @method static void nullOrIsArray($value, $message = '')
@@ -68,6 +69,12 @@ use Traversable;
  * @method static void nullOrClassExists($value, $message = '')
  * @method static void nullOrSubclassOf($value, $class, $message = '')
  * @method static void nullOrImplementsInterface($value, $interface, $message = '')
+ * @method static void nullOrPropertyExists($value, $property, $message = '')
+ * @method static void nullOrPropertyNotExists($value, $property, $message = '')
+ * @method static void nullOrMethodExists($value, $method, $message = '')
+ * @method static void nullOrMethodNotExists($value, $method, $message = '')
+ * @method static void nullOrKeyExists($value, $key, $message = '')
+ * @method static void nullOrKeyNotExists($value, $key, $message = '')
  * @method static void allString($values, $message = '')
  * @method static void allStringNotEmpty($values, $message = '')
  * @method static void allInteger($values, $message = '')
@@ -76,6 +83,7 @@ use Traversable;
  * @method static void allNumeric($values, $message = '')
  * @method static void allBoolean($values, $message = '')
  * @method static void allScalar($values, $message = '')
+ * @method static void allObject($values, $message = '')
  * @method static void allResource($values, $type = null, $message = '')
  * @method static void allIsCallable($values, $message = '')
  * @method static void allIsArray($values, $message = '')
@@ -120,6 +128,12 @@ use Traversable;
  * @method static void allClassExists($values, $message = '')
  * @method static void allSubclassOf($values, $class, $message = '')
  * @method static void allImplementsInterface($values, $interface, $message = '')
+ * @method static void allPropertyExists($values, $property, $message = '')
+ * @method static void allPropertyNotExists($values, $property, $message = '')
+ * @method static void allMethodExists($values, $method, $message = '')
+ * @method static void allMethodNotExists($values, $method, $message = '')
+ * @method static void allKeyExists($values, $key, $message = '')
+ * @method static void allKeyNotExists($values, $key, $message = '')
  *
  * @since  1.0
  *
@@ -198,6 +212,16 @@ class Assert
         if (!is_scalar($value)) {
             throw new InvalidArgumentException(sprintf(
                 $message ?: 'Expected a scalar. Got: %s',
+                self::typeToString($value)
+            ));
+        }
+    }
+
+    public static function object($value, $message = '')
+    {
+        if (!is_object($value)) {
+            throw new InvalidArgumentException(sprintf(
+                $message ?: 'Expected an object. Got: %s',
                 self::typeToString($value)
             ));
         }
@@ -709,6 +733,46 @@ class Assert
                 $message ?: 'Expected an implementation of %2$s. Got: %s',
                 self::valueToString($value),
                 self::valueToString($interface)
+            ));
+        }
+    }
+
+    public static function propertyExists($classOrObject, $property, $message = '')
+    {
+        if (!property_exists($classOrObject, $property)) {
+            throw new InvalidArgumentException(sprintf(
+                $message ?: 'Expected the property %s to exist.',
+                self::valueToString($property)
+            ));
+        }
+    }
+
+    public static function propertyNotExists($classOrObject, $property, $message = '')
+    {
+        if (property_exists($classOrObject, $property)) {
+            throw new InvalidArgumentException(sprintf(
+                $message ?: 'Expected the property %s to not exist.',
+                self::valueToString($property)
+            ));
+        }
+    }
+
+    public static function methodExists($classOrObject, $method, $message = '')
+    {
+        if (!method_exists($classOrObject, $method)) {
+            throw new InvalidArgumentException(sprintf(
+                $message ?: 'Expected the method %s to exist.',
+                self::valueToString($method)
+            ));
+        }
+    }
+
+    public static function methodNotExists($classOrObject, $method, $message = '')
+    {
+        if (method_exists($classOrObject, $method)) {
+            throw new InvalidArgumentException(sprintf(
+                $message ?: 'Expected the method %s to not exist.',
+                self::valueToString($method)
             ));
         }
     }